backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.2",
    "uvicorn>=0.40.0",
]
